requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Computer Vision & Face Detection
opencv-python>=4.5.0
Pillow>=8.3.0
mtcnn>=0.1.1

# Deep Learning & Facial Recognition
tensorflow>=2.8.0
keras-facenet>=0.3.2
deepface>=0.0.75

# Alternative face detection (optional)
# dlib>=19.22.0  # Uncomment if you want to use dlib

# Data Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Web scraping (optional, for data collection)
requests>=2.26.0
beautifulsoup4>=4.10.0

# Utilities
tqdm>=4.62.0
scikit-learn>=1.0.0

# Jupyter for exploration
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing (optional)
pytest>=6.2.0
pytest-cov>=3.0.0
